LPS-6506 LPS-5852 fixed ClassCastException
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@42492 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portlet/messageboards/service/persistence/MBMessageFinderImpl.java
@@ -90,10 +90,10 @@ public class MBMessageFinderImpl
 			qPos.add(createDate_TS);
 			qPos.add(threadId);
 
-			Iterator<Long> itr = q.list().iterator();
+			Iterator<Integer> itr = q.list().iterator();
 
 			if (itr.hasNext()) {
-				Long count = itr.next();
+				Integer count = itr.next();
 
 				if (count != null) {
 					return count.intValue();
